Moved occurrence logic into occurrence model and change some tabs for spaces
