+++ gestor_absencies/common/utils.py
@@ -1,13 +1,14 @@
-from gestor_absencies.models import Worker
 import dateutil
 
+
 def jwt_response_token_user_id(token, user=None, request=None):
-	user = Worker.objects.filter(username=request.data['username']).first()
-	return {
-		'user_id': user.id,
-		'is_admin': user.is_superuser,
-		'token': token,
-	}
+    from gestor_absencies.models import Worker
+    user = Worker.objects.filter(username=request.data['username']).first()
+    return {
+        'user_id': user.id,
+        'is_admin': user.is_superuser,
+        'token': token,
+    }
 
 
 def computable_days_between_dates(start_time, end_time, dates_types):
@@ -17,6 +18,7 @@ def computable_days_between_dates(start_time, end_time, dates_types):
         freq=dateutil.rrule.DAILY,
         byweekday=dates_types
     )))
+
 
 def find_concurrence_dates(o1, o2):	# TODO with interval
     if o1.start_time > o2.start_time and o1.start_time < o2.end_time:
@@ -33,5 +35,6 @@ def find_concurrence_dates(o1, o2):	# TODO with interval
         return o1.start_time, o1.end_time
     return o1.start_time, o1.end_time
 
+
 def between_dates(start_interval, end_interval, date):
-	return start_interval < date and end_interval > date
+    return start_interval < date and end_interval > date
